Allow mapping of segments via idpaths
At present yml files can describe a segment which allows for the
labelling of components connected to a USB tree - e.g. a hub and
connected devices. yml files can also be used to associate a segment
with a location in the USB tree (e.g. this hub is here).
The association is based on a port path (e.g. 1-5.2), however this
may change over time based on enumeration order.

Let's overcome this by allowing for associations to be made via UDEV
idpaths. These are specific to the physical hardware topology.